% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetBoxIndices}
\alias{GetBoxIndices}
\title{Fast Implementation of finding x and y order numbers to feed into the linear congruence equation.
This solves equation solves for a_i in the HIP paper.
This is essentially an internal function and shouldn't be worried about.}
\usage{
GetBoxIndices(lxy, base, J)
}
\arguments{
\item{lxy}{A Matrix of lower x y coordinates of a Halton Box inside the unit box.}

\item{base}{Co-prime Base but generally for BAS work it is 2 or 3.}

\item{J}{Integer of 2 values that represent the numbers 2^J1, 3^J2.}
}
\description{
Fast Implementation of finding x and y order numbers to feed into the linear congruence equation.
This solves equation solves for a_i in the HIP paper.
This is essentially an internal function and shouldn't be worried about.
}
