% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HaltonSeq}
\alias{HaltonSeq}
\title{Draw Halton Sequence values for a single dimension.
Note that this was borrowed from the Internet and is not my implmementation.}
\usage{
HaltonSeq(k, base, n)
}
\arguments{
\item{base}{Co-prime Base but generally for BAS work it is 2 or 3.}

\item{n}{Number of samples to draw.}

\item{x}{An integer for the starting index k >= 0.}
}
\description{
Draw Halton Sequence values for a single dimension.
Note that this was borrowed from the Internet and is not my implmementation.
}
\examples{
HaltonSeq(k = 0, base = 2, n = 10)

}
