% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterSample.r
\name{lineSamp}
\alias{lineSamp}
\title{Generate samples in linear features based on BAS mastersample}
\usage{
lineSamp(n = 10, x, seed = 0, halt = TRUE)
}
\description{
Generate samples in linear features based on BAS mastersample
}
