% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketransects.R
\name{makeTransect}
\alias{makeTransect}
\title{Make transects of a specified distance from SpatialPointsDataFrames}
\usage{
makeTransect(x, tran.length)
}
\description{
Make transects of a specified distance from SpatialPointsDataFrames
}
