% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterSample.r
\name{masterSample}
\alias{masterSample}
\title{Generate sample points in New Zealand using BAS master sample}
\usage{
masterSample(island = "South", shp, N = 100, J = c(0, 0))
}
\description{
Generates BAS sample points in a specified sample frame based on New Zealand terrestrial mastersample. Users need to specify 'island' the sample frame is in and how many points are required.
}
