% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterSample.r
\name{shape2Frame}
\alias{shape2Frame}
\title{Wrap a Halton Frame over the sample shape}
\usage{
shape2Frame(shp, bb = NULL, base = c(2, 3), J = c(2, 2), projstring = NULL)
}
\description{
Wrap a Halton Frame over the sample shape
}
