% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SolveCongruence}
\alias{SolveCongruence}
\title{Solve system of linear congruence from HIP paper to order HIP boxes.
See page 5 of Robertson et al. 2018 Halton Iterative Partitioning.
This is essentially an internal function and shouldn't be worried about.}
\usage{
SolveCongruence(A, base, J)
}
\arguments{
\item{A}{Matrix that is in numeric for computational reasons but is the a_i solutions for all HIP boxes.}

\item{base}{Co-prime Base but generally for BAS work it is 2 or 3.}

\item{J}{Integer of 2 values that represent the numbers 2^J1, 3^J2.}
}
\description{
Solve system of linear congruence from HIP paper to order HIP boxes.
See page 5 of Robertson et al. 2018 Halton Iterative Partitioning.
This is essentially an internal function and shouldn't be worried about.
}
