% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketransects.R
\name{maketransect}
\alias{maketransect}
\title{Make transects of a specified distance from SpatialPointsDataFrames}
\usage{
maketransect(x, tran.length)
}
\description{
Make transects of a specified distance from SpatialPointsDataFrames
}
